proper structure, fix endpoint discrepancies
