+++ listen-kit/src/data/mod.rs
@@ -3,7 +3,6 @@ use rig_tool_macro::tool;
 use serde::{Deserialize, Serialize};
 
 pub mod twitter;
-mod twitter_types;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Candlestick {
